Extract build from force_release target to script
This moves the conditional build dependency installation logic and
build logic from the force_release Makefile target to a shell
script build-release.sh, which force_release calls. The code is
changed to clean it up, and also to account for differences between
how output is displayed and errors reported in Makefiles and shell
scripts. (As in check-version.sh, the .sh suffix does not signify
anything about how the script is to be used: like the other shell
scripts in the project, this should be executed, no sourced.)