Fix URLs that were redirecting to another license
All the opensource.org BSD license URLs at the top of source code
files in this project had originally pointed to a page on the
3-clause BSD license that this project used and continues to use.

But over time the site was apparently reorganized and the link
became a redirect to the page about the 2-clause BSD license.
Because it is identified only as the "BSD license" in the comments
in this project that contain the links, this unfortunately makes it
so those top-of-file comments all wrongly claim that the project is
2-clause BSD licensed.

This fixes the links by replacing them with the current URL of the
opensource.org page on the 3-clause BSD license. The current URL
contains "bsd-3-clause" in it, so this specific problem is unlikely
to recur with that URL (and even if it did, the text "bsd-3-clause
is information that may clue readers in to what is going on).