Assorted small fixes/improvements to root dir docs
This contains misc. formatting fixes and minor proofreading in the
top-level documentation files, plus making the text in the LICENSE
section of README.md have links both to external information about
the license and to the license file itself.

Note that the changes to the license file are just removal of
excess whitespace (the extra blank line at the end, and spaces
appearing at the end of lines).