Omit py_modules in setup
This removes the py_modules keyword argument in the call to setup,
and further shortens/simplifies setup.py by removing the now-unused
build_py_modules function.

The packages keyword argument already covers this, because we have
no loose modules that are included in the distribution, and the
call to find_packages:

- Omits everything in test/ because it is directed to do so in the
  call.

- Omits the gitdb/ directory (currently existing as a git submodule,
  not to be confused with Python submodules), because the ext/
  directory that contains it does not itself directly contain an
  __init__.py file, so it is not a traditional package, yet ext/
  is contained and found inside the directory git/ that *is* a
  traditional package, so the ext/ directory is not a namespace
  package either.

- Includes all other modules, recursively, because they are all
  in a recursive traditional package structure under git/ that
  find_packages recognizes.

To verify that this includes the same files in the built wheel and
sdist distributions, I have listed the contents of the wheel with
"unzip -l" and the sdist .tar.gz file with "tar tf" both before and
after this change, verifying they list all the same entries.