Don't track code coverage temporary files
While running tests with coverage enabled, files of the form
.coverage.MachineName.####.###### are created temporarily. When
tests complete normally (whether or not there are failures), these
files are automatically removed. However, when tests are cancelled
with SIGINT (Ctrl+C), they are sometimes retained.

This adds another pattern to .gitignore so that, in addition to not
tracking the .coverage file that is retained after tests, these
other temporary files are also not tracked.