Update flake8 additional dependencies, fix warning
This bumps the versions of the flake8 plugins specified with pinned
versions as additional dependencies of flake8 for pre-commit.

Doing so gains a warning about a call to warnings.warn with no
stacklevel argument. This appears to be the uncommon case where the
implifit effect of stacklevel=1 is intended, so I have made that
explicit, which clarifies this intent and dismisses the warning.