Refactor "finally" cleanup in tests, fix minor bug
This refactors code in the test suite that uses try-finally, to
simplify and clarify what it is doing. An exception to this being
a refactoring is that a possible bug is fixed where a throwaway
environment variable "FOO" was patched for a test and never
unpatched.

There are two patterns refactored here:

- try-(try-except)-finally to try-except-finally. When the entire
  suite of the try-block in try-finally is itself a try-except,
  that has the same effect as try-except-finally. (Python always
  attempts to run the finally suite in either case, and does so
  after any applicable except suite is run.)

- Replacing try-finally with an appropriate context manager.

(These changes do not fix, nor introduce, any flake8 errors, but
they were found by manual search for possible problems related to
recent flake8 findings but outside of what flake8 can catch. To
limit scope, this only refactors try-finally in the test suite.)