Tweak rollback logic in log.to_file
This modifies the exception handling in log.to_file so it catches
BaseException rather than Exception and rolls back. Ordinarily we
do not want to catch BaseException, since this means catching
things like SystemExit, KeyboardInterupt, etc., but the other cases
of rolling back with LockedFD do it that strongly (both before when
try-finally was used with a flag, and now with try-except catching
BaseException to roll back the temporary-file write and reraise).
Having this behave subtly different does not appear intentional.

(This is also closer to what will happen if LockedFD becomes a
context manager and these pieces of code use it in a
with-statement: even exceptions not inheriting from Exception will
cause __exit__ to be called.)