Note that the init script can be run from Git Bash
This is probably the *only* way anyone should run that script on
Windows, but I don't know of specific bad things that happen if it
is run in some other way, such as with WSL bash, aside from messing
up line endings, which users are likely to notice anyway.

This commit also clarifies the instructions by breaking up another
paragraph that really represented two separate steps.