Drop claim about Cygwin not having git-daemon
On a current Cygwin system with git 2.39.0 (the latest version
offered by the Cygwin package manager), git-daemon is present, with
the Cygwin path /usr/libexec/git-core/git-daemon.exe.

In addition, the cygwin-test.yml workflow does not take any special
steps to allow git-daemon to work, but all tests pass in it even
without skipping or xfailing tests that seem related to git-daemon.

The git_daemon_launched function in test/lib/helper.py only invokes
git-daemon directly (rather than through "git daemon") when is_win
evaluates true, which only happens on native Windows systems, not
Cygwin, which is treated the same as (other) Unix-like systems and
still works. So maybe Cygwin git-daemon was never a special case.

Whether or not it was, the message about git-daemon needing to be
findable in a PATH search is also under an is_win check, and thus
is never shown on Cygwin. So I've removed the Cygwin part of that
message. (Because the path shown is a MinGW-style path, I have kept
the wording about that being for MinGW-git, even though it is no
longer needed to disambiguate the Cygwin case.)