Add a missing PermissionError xfail on Windows
One of the tests that was commented as being skipped as a result of
SkipTest rasied in git.util.rmtree or one of the functions that
calls it, test_git_submodule_compatibility, was not skipped in that
way and was actually failing on Windows with PermissionError. It
appears that the cause of failure changed over time, so that it
once involved rmtree but no longer does.

This removes the outdated comment and adds an xfail mark instead,
specific to PermissionError and with a message identifying where in
the test case's logic the PermissionError is currently triggered.