Update "ACTUALLY skipped by" comments
The remaining "ACTUALLY skipped by" comments in the test suite were
for tests that are actually skipped by SkipTest exceptions raised
from the code under test. But the information provided about where
in the code they were skipped was out of date, and also not
detailed enough because references to line numbers become stale
when code is added or removed in the referenced module before the
referenced code.

This updates them and also provides more detailed information about
the referenced code doing the skipping.

The error messages are the same as before, and the paths are the
same in relevant details, so this doesn't modify those parts of the
comments.